fix : 트러블 슈팅 기능 버그 수정
1. 트러블 상세 기능 덧글 및 태그 안가져오는 버그 해결
2. 덧글 수정 안되는 부분 해결
3. 덧글 삭제 안되는 부분 해결
4. 트러블슈팅 수정 안되는 부분 해결
5. 트러블 슈팅 삭제 안되는 부분 해결
6. 태그 수정 안되는 부분 해결

231105 suhyun

+++ troubleShooting/src/main/java/com/orientalSalad/troubleShot/tag/serivice/TagService.java
@@ -12,9 +12,11 @@ import com.orientalSalad.troubleShot.tag.repository.TagRepository;
 import com.orientalSalad.troubleShot.tag.repository.TroubleTagRepository;
 
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 
 @Service
 @RequiredArgsConstructor
+@Slf4j
 public class TagService {
 	private final TagRepository tagRepository;
 	private final TroubleTagRepository troubleTagRepository;
@@ -22,7 +24,9 @@ public class TagService {
 	//트러블 슈팅에 태그를 다는 메소드
 	public boolean deleteTag(Long troubleShootingSeq){
 		List<TroubleTagEntity> tagEntityList = troubleTagRepository.findByTroubleSeq(troubleShootingSeq);
-
+		if(tagEntityList == null){
+			return true;
+		}
 		for(TroubleTagEntity troubleTagEntity : tagEntityList){
 			troubleTagRepository.delete(troubleTagEntity);
 		}
@@ -31,6 +35,9 @@ public class TagService {
 	}
 	//트러블 슈팅에 태그를 다는 메소드
 	public boolean attachTag(List<String> tagDTOList, Long troubleShootingSeq){
+		if(tagDTOList == null){
+			return true;
+		}
 		for(String tagName : tagDTOList){
 			TagEntity tagEntity = tagRepository.findByName(tagName);
 
@@ -59,11 +66,17 @@ public class TagService {
 		//기존 태그 삭제
 		deleteTag(troubleShootingSeq);
 		//새로운 태그 달기
+		if(tagDTOList == null){
+			log.info("새롭게 달 태그가 없습니다.");
+			return true;
+		}
 		for(String tagName : tagDTOList){
+			log.info("새로운 태그 : "+tagName+" with "+troubleShootingSeq);
 			TagEntity tagEntity = tagRepository.findByName(tagName);
 
 			//없는 태그인 경우 생성
 			if(tagEntity == null){
+				log.info("없는 태그");
 				TagDTO tagDTO = TagDTO.builder()
 					.name(tagName)
 					.build();
@@ -76,7 +89,8 @@ public class TagService {
 				= troubleTagRepository.findByTroubleSeqAndTagSeq(troubleShootingSeq,tagEntity.getSeq());
 
 			//트러블 슈팅 - 태그 관계 생성
-			if(troubleTagEntity != null){
+			if(troubleTagEntity == null){
+				log.info("태그 연결");
 				troubleTagEntity = TroubleTagEntity.builder()
 					.tagSeq(tagEntity.getSeq())
 					.troubleSeq(troubleShootingSeq)
